Final cleanup.
1. replace params_spec.py with params_spec_v2.py (i.e. rename params_spec_v2.py to params_spec.py, removing the original params_spec.py).

2. dcnm_image_policy.py: Change import ParamsSpec() from the new params_spec.

3. payload.py: remove debug statements.

4.
